+++ tests/profiles/universal_test.py
@@ -795,7 +795,8 @@ def test_check_unwanted_tables():
     "TSI2",
     "TSI3",
     "TSI5",
-    "prop" # FIXME: Why is this one unwanted?
+    "prop", # FIXME: Why is this one unwanted?
+    "MVAR", # Bugs in DirectWrite
   ]
   # Our reference Mada Regular font is good here:
   ttFont = TTFont(TEST_FILE("mada/Mada-Regular.ttf"))